+++ tensorflow/compiler/xla/mlir_hlo/tosa/lib/Transforms/legalize_mhlo.cc
@@ -220,7 +220,7 @@ struct ConvertMhloIotaOp : public OpRewritePattern<mhlo::IotaOp> {
     size_t tileMultiplesSize = resultShape.size();
     tileMultiples.resize(tileMultiplesSize);
 
-    for (int i = 0; i < tileMultiplesSize; i++) {
+    for (size_t i = 0; i < tileMultiplesSize; i++) {
       if (i == iotaDimension) {
         tileMultiples[i] = 1;
       } else {
@@ -323,7 +323,7 @@ struct ConvertMhloSliceOp : public OpRewritePattern<mhlo::SliceOp> {
     llvm::SmallVector<int64_t, 2> startIndicesI64;
     startIndicesI64.resize(startIndices.size());
 
-    for (int64_t i = 0; i < startIndices.size(); i++) {
+    for (int64_t i = 0; i < static_cast<int64_t>(startIndices.size()); i++) {
       size[i] = endIndices[i] - startIndices[i];
       startIndicesI64[i] = startIndices[i];
     }
